python_backend/requirements.txt
fastapi
uvicorn[standard]
openai
motor
python-dotenv
pydantic
pydantic-settings
python-multipart
Pillow
img2pdf
httpx